[Silver III] Title: 주유소, Time: 776 ms, Memory: 98048 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -5,10 +5,9 @@ public class Main {
 
 	public static void main(String[] args) {
 	Scanner sc= new Scanner(System.in);
-		
-		int N = Integer.parseInt(sc.next());
-		int [] a = new int[N-1];
-		int [] b = new int[N];
+		long N = Long.parseLong(sc.next());
+		long [] a = new long[(int) (N-1)];
+		long [] b = new long[(int) N];
 		
 		
 		for(int i= 0 ; i <a.length ; i++) {
@@ -17,8 +16,8 @@ public class Main {
 		for(int i= 0 ; i <b.length ; i++) {
 			b[i] = Integer.parseInt(sc.next());
 		}
-		int min = b[0] ;
-		int sum = 0 ;
+		long min = b[0] ;
+		long sum = 0 ;
 		for(int i= 0 ; i < b.length ; i++) {
 			if( i == b.length -1) {
 				break;
